Modularized the files and flow, coderefactor
