+++ elfparser/JSONTestSuite/test_parsing/n_structure_object_followed_by_closing_object.json
@@ -1,1 +1,1 @@
-{}}
+p{}}
